Replace forward reference with Fwd.h includes for HcalDigi
This is to support the EVOLUTION branch
