+++ DataFormats/HcalDigi/interface/HBHEDataFrameFwd.h
@@ -1,0 +1,6 @@
+#ifndef DataFormats_HcalDigi_HBHEDataFrameFwd_h
+#define DataFormats_HcalDigi_HBHEDataFrameFwd_h
+
+class HBHEDataFrame;
+
+#endif
