+++ DataFormats/HcalDigi/interface/HFDataFrameFwd.h
@@ -1,0 +1,6 @@
+#ifndef DataFormats_HcalDigi_HFDataFrameFwd_h
+#define DataFormats_HcalDigi_HFDataFrameFwd_h
+
+class HFDataFrame;
+
+#endif
