+++ DataFormats/HcalDigi/interface/HODataFrameFwd.h
@@ -1,0 +1,6 @@
+#ifndef DataFormats_HcalDigi_HODataFrameFwd_h
+#define DataFormats_HcalDigi_HODataFrameFwd_h
+
+class HODataFrame;
+
+#endif
