+++ DataFormats/HcalDigi/interface/ZDCDataFrameFwd.h
@@ -1,0 +1,6 @@
+#ifndef DataFormats_HcalDigi_ZDCDataFrameFwd_h
+#define DataFormats_HcalDigi_ZDCDataFrameFwd_h
+
+class ZDCDataFrame;
+
+#endif
